[@foal/acceptance-tests] Rename tests.
