+++ packages/acceptance-tests/src/404-page.spec.ts
@@ -4,7 +4,8 @@ import * as request from 'supertest';
 // FoalTS
 import { createApp, Get, HttpResponseNotFound, HttpResponseOK } from '@foal/core';
 
-it('404', async () => {
+it('404 page: users should be able to define a controller method that catches requests'
+  + ' that have no handler.', async () => {
 
   class ViewController {
     @Get('/home')
